Move submissions list from all teams page to specific team page
