+++ src/pages/Competitions/CompetitionTeamPages/AllTeamsPage/index.tsx
@@ -2,47 +2,9 @@ import React, { useEffect, useState, useContext } from 'react';
 import DefaultLayout from '../../../../components/layouts/default';
 import './index.less';
 import { useParams, Link } from 'react-router-dom';
-import { getRegisteredState, getTeams, getTeamInfo, getSubmissionDetails } from '../../../../actions/teams/utils';
+import { getRegisteredState, getTeams, getTeamInfo } from '../../../../actions/teams/utils';
 import UserContext from '../../../../UserContext';
 import BackLink from '../../../../components/BackLink';
-
-import { Table, Button, Modal } from 'antd';
-import { ColumnsType } from 'antd/lib/table';
-
-interface SubmissionData {
-  description: string;
-  score: number;
-  tags: string[];
-  submissionDate: string;
-  // last: Date;
-}
-
-const columns: ColumnsType<SubmissionData> = [
-  {
-    title: 'Submission Date',
-    dataIndex: 'submissionDate',
-    // defaultSortOrder: 'descend',
-    // sorter: (a, b) => a.score - b.score,
-  },
-  {
-    title: 'Description',
-    dataIndex: 'description',
-    // defaultSortOrder: 'descend',
-    // sorter: (a, b) => a.score - b.score,
-  },
-  {
-    title: 'Tags',
-    dataIndex: 'tags',
-    // defaultSortOrder: 'descend',
-    // sorter: (a, b) => a.score - b.score,
-  },
-  {
-    title: 'Score',
-    dataIndex: 'score',
-    // defaultSortOrder: 'descend',
-    sorter: (a, b) => a.score - b.score,
-  },
-];
 
 // Block for each team
 const Team = (team: any) => {
@@ -67,10 +29,6 @@ const CompetitionAllTeamsPage = () => {
   let { competitionName } = useParams<{ competitionName: string }>();
   const username = "testinguser4";
 
-  // submission page messing around 
-  const [submissionData, setSubmissionData] = useState<SubmissionData[]>([]);
-  const [submissionIds, setSubmissionIds] = useState<any>([]);
-
   // Get all team data
   useEffect(() => {
     if (user.loggedIn) {
@@ -78,42 +36,16 @@ const CompetitionAllTeamsPage = () => {
         setIsRegistered(res.data.registered);
       })
     }
-    getTeams(competitionName).then((res) => {
-      setTeams(res.data);
-    });
   }, []);
 
-  // Store submission IDs (right now, of random IDs)
   useEffect(() => {
     if (isRegistered) {
-      getTeamInfo(competitionName, "testTeam1").then((res) => {
-        // console.log(res.data.submitHistory);
-
-        // These are just random IDs
-        setSubmissionIds([
-          "63c396f9671b14068b17f681",
-          "63c396f9671b14068b17f682",
-          "63c396f9671b14068b17f683"
-        ]);
+      getTeams(competitionName).then((res) => {
+        setTeams(res.data);
       });
     }
-  }, [isRegistered])
 
-  // Get submission data from submission IDs
-  useEffect(() => {
-    submissionIds.map((id: any) => {
-      getSubmissionDetails(competitionName, id).then((res) => {
-        let submission = res.data[0];
-        let submissionDetails = {
-          submissionDate: submission.submissionDate,
-          description: submission.description,
-          tags: submission.tags,
-          score: submission.score
-        }
-        setSubmissionData(submissionData => [...submissionData, submissionDetails]);
-      })
-    })
-  }, [submissionIds])
+  }, [isRegistered])
 
   return (
     <DefaultLayout>
@@ -123,8 +55,6 @@ const CompetitionAllTeamsPage = () => {
         {isRegistered ? (
           <div>
             <div className='main-section'>
-              <h2>My Team Submissions</h2>
-              <Table columns={columns} dataSource={submissionData} />
               <h2>Teams</h2>
               {teams.map((team: any) => {
                 return (Team(team));
