+++ src/pages/Competitions/CompetitionTeamPages/SpecificTeamPage/index.tsx
@@ -2,19 +2,54 @@ import React, { useEffect, useState, useContext } from 'react';
 import DefaultLayout from '../../../../components/layouts/default';
 import './index.less';
 import { useParams } from 'react-router-dom';
-import { getRegisteredState, getTeamInfo } from '../../../../actions/teams/utils';
+import { getRegisteredState, getTeamInfo, getSubmissionDetails } from '../../../../actions/teams/utils';
 import UserContext from '../../../../UserContext';
 import BackLink from '../../../../components/BackLink';
+import { Table } from 'antd';
+import { ColumnsType } from 'antd/lib/table';
+
+interface SubmissionData {
+  description: string;
+  score: number;
+  tags: string[];
+  submissionDate: string;
+}
+
+const columns: ColumnsType<SubmissionData> = [
+  {
+    title: 'Submission Date',
+    dataIndex: 'submissionDate',
+    defaultSortOrder: 'descend',
+    // sorter: (a, b) => a.score - b.score,
+  },
+  {
+    title: 'Description',
+    dataIndex: 'description',
+  },
+  {
+    title: 'Tags',
+    dataIndex: 'tags',
+  },
+  {
+    title: 'Score',
+    dataIndex: 'score',
+    sorter: (a, b) => a.score - b.score,
+  },
+];
 
 const CompetitionSpecificTeamPage = () => {
 
   const { user } = useContext(UserContext);
-  const [isRegistered, setIsRegistered] = useState<any>(false);
-  const [team, setTeam] = useState<any>({});
+  const [isRegistered, setIsRegistered] = useState<boolean>(false);
+  const [teamInfo, setTeamInfo] = useState<any>({});
   const [teamMembers, setTeamMembers] = useState<any>([]);
-
+  const [isMyTeam, setIsMyTeam] = useState<boolean>(false);
+  const [submissionData, setSubmissionData] = useState<SubmissionData[]>([]);
+  const [submissionIds, setSubmissionIds] = useState<any>([]);
   let { competitionName, teamName } = useParams<{ competitionName: string, teamName: string }>();
   const username = "testinguser4"; // replace with user.username
+
+  // Check if user is logged in
   useEffect(() => {
     if (user.loggedIn) {
       // TODO: it's hardcoded to "testinguser1"; change it to user.username
@@ -22,12 +57,49 @@ const CompetitionSpecificTeamPage = () => {
         setIsRegistered(res.data.registered);
       })
     }
-
-    getTeamInfo(competitionName, teamName).then((res) => {
-      setTeam(res.data);
-      setTeamMembers(res.data.teamMembers);
-    });
   }, []);
+
+  // Get info on team if user is registered in competition
+  useEffect(() => {
+    if (isRegistered) {
+      getTeamInfo(competitionName, teamName).then((res) => {
+        setTeamInfo(res.data);
+        setTeamMembers(res.data.teamMembers);
+      });
+    }
+  }, [isRegistered])
+
+  // Retrieve submission IDs if user is in their team's page
+  useEffect(() => {
+    if (teamMembers.length > 0 && teamMembers.includes(username)) {
+      setIsMyTeam(true);
+
+      // For real thing, use this code instead:
+      // setSubmissionIds(teamInfo.submitHistory);
+      // These are just random submission IDs to show that the table works
+      setSubmissionIds([
+        "63c396f9671b14068b17f681",
+        "63c396f9671b14068b17f682",
+        "63c396f9671b14068b17f683"
+      ]);
+    }
+  }, [teamMembers])
+
+  // Get submission data from submission IDs
+  useEffect(() => {
+    submissionIds.map((id: any) => {
+      getSubmissionDetails(competitionName, id).then((res) => {
+        let submission = res.data[0];
+        let submissionDetails = {
+          submissionDate: submission.submissionDate,
+          description: submission.description,
+          tags: submission.tags,
+          score: submission.score
+        }
+        setSubmissionData(submissionData => [...submissionData, submissionDetails]);
+      })
+    })
+  }, [submissionIds])
 
   return (
     <DefaultLayout>
@@ -36,15 +108,14 @@ const CompetitionSpecificTeamPage = () => {
         <BackLink to="../" />
         {isRegistered ? (
           <div className='main-section'>
-            <h2>Team {team.teamName}</h2>
-            {/* TODO: Change testinguser1 to user.username */}
-            {teamMembers.includes(username) &&
+            <h2>Team {teamInfo.teamName}</h2>
+            {isMyTeam &&
               <div className='block'>
                 <p>Invite your friends to join this team!</p>
-                <h4>Join Code: {team.joinCode}</h4>
+                <h4>Join Code: {teamInfo.joinCode}</h4>
               </div>
             }
-            <h3><span className="subheader">Best Score: {team.bestScore}</span></h3>
+            <h3><span className="subheader">Best Score: {teamInfo.bestScore}</span></h3>
 
             <h3><span className="subheader">Members</span></h3>
             <div>
@@ -54,7 +125,14 @@ const CompetitionSpecificTeamPage = () => {
             </div>
 
             <h3><span className="subheader">About</span></h3>
-            <p>{team.teamDescription}</p>
+            <p>{teamInfo.teamDescription}</p>
+
+            {isMyTeam &&
+              <>
+                <h3>Submissions</h3>
+                <Table columns={columns} dataSource={submissionData} />
+              </>
+            }
           </div>
         ):(
           <p className='errorMessage'>
